+++ programs/stake/src/lib.rs
@@ -12,6 +12,9 @@ use solana_sdk::{
 };
 
 pub mod config;
+pub mod points;
+#[doc(hidden)]
+pub mod rewards;
 pub mod stake_instruction;
 pub mod stake_state;
